A first version of the GUI
A very simple version of the GUI with a window, a menu, some button. All
without functionality.
