+++ POI/ViAn/main.cpp
@@ -1,6 +1,13 @@
 #include "mainwindow.h"
 #include <QApplication>
 
+/**
+ * @brief qMain
+ * Constructor
+ * @param argc an int variable
+ * @param argv a list of chars
+ * @return exits the program
+ */
 int main(int argc, char *argv[])
 {
     QApplication a(argc, argv);
